requirements.txt
pygame-ce>=2.5.0
tomli >= 1.1.0 ; python_version < "3.11"